backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic-settings==2.7.0
python-multipart==0.0.12
httpx==0.28.1
pytest==8.3.4
alembic==1.14.1
